Implemented (unfoturnately truncated) exporting to CSV.
Need to see if there's a (simple) method to have the dictionaries within the dict_of_dicts side-by-side in a CSV. Else, could always have them in different CSVs instead.
Currently, all dicts are exported to a single CSV, which apparently has an upper limit on the number of rows and ends up truncating values past row ~100000.