Added option to choose alternative directory for images.
Also removed PIL max image size warning. This, however, does *not* allow arbitrary filesize images to be opened... Will need to be scaled down or for a Text Image Reader class to be implemented.
Also put skeleton in for an on-disk TextImageReader (as loading ~2GB .txt files in memory is not the fastest thing in the world...).